+++ src/graph_onedrive/_cli.py
@@ -1,19 +1,24 @@
 """Command line interface tools for the Python package Graph-OneDrive.
 Run terminal command 'graph-onedrive --help' or 'python -m graph-onedrive --help' for details.
 """
+from __future__ import annotations
+
 import argparse
 import json
 import os
 from datetime import datetime
-from typing import Optional
 from typing import Sequence
-from typing import Tuple
 
 import graph_onedrive._main as graph_onedrive
 from graph_onedrive.__init__ import __version__
 
 
-def main(argv: Optional[Sequence[str]] = None) -> int:
+CONFIG_EXT = ".json"
+CONFIG_DEF_FILE = "config" + CONFIG_EXT
+CONFIG_DEF_KEY = "onedrive"
+
+
+def main(argv: Sequence[str] | None = None) -> int:
     """Command line interface tools for the Python package Graph OneDrive.
     Use command --help for details.
     """
@@ -67,8 +72,8 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
     if not (args.configure or args.authenticate or args.instance):
         parser.error("No action provided, use --help for details")
 
-    if args.file and not args.file.endswith(".json"):
-        parser.error("--file path must end in a json file")
+    if args.file and not args.file.endswith(CONFIG_EXT):
+        parser.error(f"--file path must have {CONFIG_EXT} extension")
 
     if args.key and args.key == "":
         parser.error("--key provided can not be blank")
@@ -85,34 +90,35 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
     return 0
 
 
-def config(config_path: Optional[str] = None, config_key: Optional[str] = None) -> None:
+def config(config_path: str | None = None, config_key: str | None = None) -> None:
     """Create a configuration file."""
 
     # Set the export directory
     if not config_path:
         if (
-            input("Save config.json in current working directory [Y/n]: ")
+            input(f"Save {CONFIG_DEF_FILE} in current working directory [Y/n]: ")
             .strip()
             .lower()
             == "n"
         ):
             config_path = input("Path to save config (including file name): ").strip()
-            if not config_path.endswith(".json"):
-                config_path = config_path + ".json"
         else:
-            config_path = os.path.join(os.getcwd(), "config.json")
+            config_path = os.path.join(os.getcwd(), CONFIG_DEF_FILE)
+
+    if not config_path.endswith(CONFIG_EXT):
+        config_path += CONFIG_EXT
 
     # Set config dictionary key
     if not config_key:
         if (
-            input("Use config dictionary key default 'onedrive' [Y/n]: ")
+            input(f"Use config dictionary key default '{CONFIG_DEF_KEY}' [Y/n]: ")
             .strip()
             .lower()
             == "n"
         ):
             config_key = input("Config dictionary key to use: ").strip()
         else:
-            config_key = "onedrive"
+            config_key = CONFIG_DEF_KEY
 
     # Load the current file if it exists, otherwise create dictionary
     if os.path.isfile(config_path):
@@ -156,28 +162,22 @@ def config(config_path: Optional[str] = None, config_key: Optional[str] = None) 
     print(f"Configuration saved to: {config_path}")
 
 
-def authenticate(
-    config_path: Optional[str] = None, config_key: Optional[str] = None
-) -> None:
+def authenticate(config_path: str | None = None, config_key: str | None = None) -> None:
     """Authenticate with OneDrive and then save the configuration to file."""
     # Get the config file path
     config_path, config_key = get_config_file(config_path, config_key)
 
     # Create the instance
-    onedrive = graph_onedrive.create_from_config_file(
-        config_path=config_path, config_key=config_key
-    )
+    onedrive = graph_onedrive.OneDrive.from_json(config_path, config_key)
 
     # Save the config
-    graph_onedrive.save_to_config_file(
-        onedrive_instance=onedrive, config_path=config_path, config_key=config_key
-    )
+    onedrive.to_json(config_path, config_key)
     print(f"Refresh token saved to configuration: {config_path}")
 
 
 def get_config_file(
-    config_path: Optional[str] = None, config_key: Optional[str] = None
-) -> Tuple[str, str]:
+    config_path: str | None = None, config_key: str | None = None
+) -> tuple[str, str]:
     """Sets a config path and key by searching the cwd with assistance from from user."""
 
     if not config_path:
@@ -186,7 +186,7 @@ def get_config_file(
         count = 0
         for root, dirs, files in os.walk(cwd_path):
             for file_name in files:
-                if file_name.endswith(".json"):
+                if file_name.endswith(CONFIG_EXT):
                     if (
                         input(f"Found: {file_name} Use this? [Y/n]: ").strip().lower()
                         != "n"
@@ -204,8 +204,8 @@ def get_config_file(
         if not config_path:
             while True:
                 config_path = input("Path to config file: ").strip()
-                if not config_path.endswith(".json"):
-                    config_path = config_path + ".json"
+                if not config_path.endswith(CONFIG_EXT):
+                    config_path = config_path + CONFIG_EXT
                 if os.path.isfile(config_path):
                     break
                 print("Path could not be validated, please try again.")
@@ -216,7 +216,7 @@ def get_config_file(
 
     # Check the config key
     if not config_key:
-        config_key = "onedrive"
+        config_key = CONFIG_DEF_KEY
         if config_key in config:
             if (
                 input("Config dictionary key 'onedrive' found. Use this key? [Y/n]: ")
@@ -232,9 +232,7 @@ def get_config_file(
     return config_path, config_key
 
 
-def instance(
-    config_path: Optional[str] = None, config_key: Optional[str] = None
-) -> None:
+def instance(config_path: str | None = None, config_key: str | None = None) -> None:
     """Interact directly with OneDrive in the command line to perform simple tasks and test the configuration."""
 
     # Check with user if using config file
@@ -253,8 +251,8 @@ def instance(
         )
 
         # Create session
-        onedrive = graph_onedrive.create_from_config_file(
-            config_path=config_path_verified, config_key=config_key_verified
+        onedrive = graph_onedrive.OneDrive.from_json(
+            config_path_verified, config_key_verified
         )
     else:
         print("Manual configuration entry:")
@@ -264,7 +262,7 @@ def instance(
         if tenant == "":
             tenant = "common"
         refresh_token = input("refresh_token (leave blank to reauthenticate): ").strip()
-        onedrive = graph_onedrive.create(
+        onedrive = graph_onedrive.OneDrive(
             client_id, client_secret, tenant, refresh_token
         )
 
@@ -298,7 +296,7 @@ def instance(
                 onedrive.get_usage(verbose=True)
 
             elif command in ["li", "list"]:
-                folder_id: Optional[str] = input(
+                folder_id: str | None = input(
                     "Folder id to look into (enter nothing for root): "
                 ).strip()
                 if folder_id == "":
@@ -357,7 +355,7 @@ def instance(
                 print(response)
 
             elif command in ["md", "mkdir"]:
-                parent_folder_id: Optional[str] = input(
+                parent_folder_id: str | None = input(
                     "Parent folder id (enter nothing for root): "
                 ).strip()
                 if parent_folder_id == "":
@@ -373,7 +371,7 @@ def instance(
                 item_id = input("Item id of the file/folder to move: ").strip()
                 new_folder_id = input("New parent folder id: ").strip()
                 if input("Specify a new file name? [y/N]: ").strip().lower() == "y":
-                    new_file_name: Optional[str] = input(
+                    new_file_name: str | None = input(
                         "New file name (with extension): "
                     ).strip()
                 else:
@@ -443,7 +441,7 @@ def instance(
                 )
                 print(f"New file item id: {item_id}")
 
-            elif command in ["od", "onedrive", "drive"]:
+            elif command in ["od", CONFIG_DEF_KEY, "drive"]:
                 print("Drive id:    ", onedrive._drive_id)
                 print("Drive name:  ", onedrive._drive_name)
                 print("Drive type:  ", onedrive._drive_type)
@@ -480,10 +478,9 @@ def instance(
 
     finally:
         if use_config_file:
-            graph_onedrive.save_to_config_file(
-                onedrive,
-                config_path=config_path_verified,
-                config_key=config_key_verified,
+            onedrive.to_json(
+                config_path_verified,
+                config_key_verified,
             )